Extract `apply_collection_config_diffs`
Building the diff list and applying it to a config is separate from turning each diff into an
action, and `TableOfContent::update_collection` needs the first half to check an operation
before it applies any of it.
